requirements.txt
anyio==4.4.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
ez_setup==0.9
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.15.4
Flask==3.0.3
Flask-Session==0.8.0
fsspec==2024.6.1
future==1.0.0
gradio_client==1.0.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
msgspec==0.18.6
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pyffmpeg==2.4.2.18.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.2.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
waitress==3.0.0
websockets==11.0.3
Werkzeug==3.0.3
